+++ fabric/src/main/java/ru/dimaskama/webcam/fabric/client/render/WebcamRenderLayer.java
@@ -1,7 +1,6 @@
 package ru.dimaskama.webcam.fabric.client.render;
 
 import com.mojang.blaze3d.vertex.PoseStack;
-import com.mojang.blaze3d.vertex.VertexConsumer;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.player.LocalPlayer;
@@ -13,7 +12,6 @@ import net.minecraft.client.renderer.entity.state.PlayerRenderState;
 import net.minecraft.util.Mth;
 import org.joml.Quaternionf;
 import org.joml.Vector2fc;
-import ru.dimaskama.webcam.config.VideoDisplayShape;
 import ru.dimaskama.webcam.fabric.client.DisplayingVideo;
 import ru.dimaskama.webcam.fabric.client.duck.PlayerRenderStateDuck;
 import ru.dimaskama.webcam.net.VideoSource;
@@ -43,11 +41,8 @@ public class WebcamRenderLayer<M extends HumanoidModel<PlayerRenderState>> exten
                             entityModel.root().translateAndRotate(poseStack);
                             entityModel.getHead().translateAndRotate(poseStack);
                             PoseStack.Pose pose = poseStack.last();
-                            VertexConsumer consumer = consumers.getBuffer(WebcamRenderTypes.square(renderData.textureId()));
-                            consumer.addVertex(pose, -0.25F, -0.5F, -0.26F).setUv(0.0F, 0.0F);
-                            consumer.addVertex(pose, -0.25F, 0.0F, -0.26F).setUv(0.0F, 1.0F);
-                            consumer.addVertex(pose, 0.25F, 0.0F, -0.26F).setUv(1.0F, 1.0F);
-                            consumer.addVertex(pose, 0.25F, -0.5F, -0.26F).setUv(1.0F, 0.0F);
+                            poseStack.translate(0.0F, -0.25F, -0.26F);
+                            WebcamRenderer.renderSquare(renderData.textureId(), pose, consumers, 0.25F, 0.25F);
                             poseStack.popPose();
                         } else if (renderData.source() instanceof VideoSource.AboveHead aboveHead) {
                             poseStack.pushPose();
@@ -58,15 +53,9 @@ public class WebcamRenderLayer<M extends HumanoidModel<PlayerRenderState>> exten
                                     (rot != null ? rot.x() : -entityRenderDispatcher.camera.getXRot()) * Mth.DEG_TO_RAD,
                                     0.0F
                             ));
-                            float sc = 0.5F * aboveHead.getSize();
+                            float halfSize = 0.5F * aboveHead.getSize();
                             PoseStack.Pose pose = poseStack.last();
-                            VertexConsumer consumer = consumers.getBuffer(aboveHead.getShape() == VideoDisplayShape.ROUND
-                                    ? WebcamRenderTypes.round(renderData.textureId())
-                                    : WebcamRenderTypes.square(renderData.textureId()));
-                            consumer.addVertex(pose, -sc, -sc, 0.0F).setUv(0.0F, 0.0F);
-                            consumer.addVertex(pose, -sc, sc, 0.0F).setUv(0.0F, 1.0F);
-                            consumer.addVertex(pose, sc, sc, 0.0F).setUv(1.0F, 1.0F);
-                            consumer.addVertex(pose, sc, -sc, 0.0F).setUv(1.0F, 0.0F);
+                            WebcamRenderer.render(renderData.textureId(), pose, consumers, halfSize, halfSize, aboveHead.getShape());
                             poseStack.popPose();
                         }
                     }